requirements.txt
asgiref==3.4.1
black==22.8.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==3.2.9
django-ckeditor==6.5.1
django-crispy-forms==1.13.0
django-environ==0.9.0
django-js-asset==2.0.0
Faker==14.2.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==8.4.0
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.12
